Added room rest webservice that looks up a room by id. A lot it is hardcoded.

+++ lms/src/main/java/dk/itu/gsd/lms/integration/exposed/building/BuildingDto.java
@@ -1,14 +1,20 @@
 package dk.itu.gsd.lms.integration.exposed.building;
 
+/**
+ * 
+ * @author HP
+ * This class represent the attributes or model we present as a webservice. 
+ * That is, if you fill out these attributes they will be exposed through the webservice.
+ */
 public class BuildingDto {
 
-	private long energyConsumptionValue;
+	private double energyConsumptionValue;
 
-	public long getEnergyConsumptionValue() {
+	public double getEnergyConsumptionValue() {
 		return energyConsumptionValue;
 	}
 
-	public void setEnergyConsumptionValue(long energyConsumptionValue) {
+	public void setEnergyConsumptionValue(double energyConsumptionValue) {
 		this.energyConsumptionValue = energyConsumptionValue;
 	}
 }
